requirements.txt
alembic==1.13.3
asyncpg==0.30.0
fastapi==0.115.4
httpx==0.27.2
psycopg2-binary==2.9.10
pydantic~=2.10.3
python-decouple==3.8
pytz==2024.2
sqlalchemy==2.0.36
uvicorn==0.22.0